src/main/java/com/rabo/statement/processor/tasklet/StepExecutionDecider.java
package com.rabo.statement.processor.tasklet;

import org.springframework.batch.core.ExitStatus;
import org.springframework.batch.core.JobExecution;
import org.springframework.batch.core.StepExecution;
import org.springframework.batch.core.job.flow.FlowExecutionStatus;
import org.springframework.batch.core.job.flow.JobExecutionDecider;
import org.springframework.stereotype.Component;

@Component
public class StepExecutionDecider implements JobExecutionDecider {

    @Override
    public FlowExecutionStatus decide(JobExecution jobExecution, StepExecution stepExecution) {
        if(stepExecution.getExitStatus().equals(ExitStatus.COMPLETED)){
            return FlowExecutionStatus.COMPLETED;
        }
        return FlowExecutionStatus.STOPPED;
    }
}
